Simplify fs::read documentation
The previous example relies on `net::SocketAddr` which is unnecessary for general `fs::read` usage. It also causes an error unless `address.txt` is in a specific format and that isn't called out in the example, it emits a warning for the unused variable, and includes a variable named `foo` which doesn't convey any meaning. This is a more basic example suitable for new rust users.